feat: UTXOTag in another file
