+++ src/Data/Models/FMUTXO.cs
@@ -92,12 +92,4 @@ namespace NodeGuard.Data.Models
             return !Equals(left, right);
         }
     }
-    
-    public class UTXOTag : Entity
-    {
-        public string Key { get; set; }
-        public string Value { get; set; }
-        public int FMUTXOId { get; set; }
-        public FMUTXO FMUTXO { get; set; }
-    }
 }
